Create record feature ok

+++ client/src/components/admin/ActRow.jsx
@@ -1,7 +1,7 @@
 import { useState } from 'react'
 import { useDispatch } from 'react-redux'
 import { BiEdit, BiTrash } from 'react-icons/bi'
-import { deleteActivity } from '../../redux/activitiesSlice'
+import { deleteActivity } from '../../redux/activitySlice'
 import { successMessage, errorMessage } from '../../utils/messages'
 import Swal from 'sweetalert2'
 import TemplateForm from '../forms/TemplateForm'
@@ -14,11 +14,8 @@ const ActRow = ({ activity }) => {
     const deleteAct = async () => {
         try {
             const response = await dispatch(deleteActivity(activity._id))
-            if (response.payload.success) {
-                successMessage('Actividad eliminada')
-            } else {
-                throw new Error('Ha ocurrido un error. Intente más tarde')
-            }
+            if (response.payload.success) successMessage('Actividad eliminada')
+            else throw new Error('Ha ocurrido un error. Intente más tarde')
         } catch (error) {
             errorMessage(error.message)
         }
@@ -27,14 +24,14 @@ const ActRow = ({ activity }) => {
     const confirmation = () => {
         Swal.fire({
             text: "¿Estás seguro/a?",
-            icon: 'warning',
+            icon: 'question',
             showCancelButton: true,
             confirmButtonText: 'Sí, borrar',
-            cancelButtonText: 'Cancelar'
+            cancelButtonText: 'Cancelar',
+            confirmButtonColor: '#e11d48',
+            cancelButtonColor: '#16a34a'
         }).then((result) => {
-            if (result.isConfirmed) {
-                deleteAct()
-            }
+            if (result.isConfirmed) deleteAct()
         })
     }
 
@@ -42,11 +39,11 @@ const ActRow = ({ activity }) => {
         <Box>
             <div className='flex justify-center gap-4 mb-4'>
                 <BiEdit
-                    className='text-3xl fill-blue-400 cursor-pointer'
+                    className='text-3xl fill-blue-400 hover:fill-blue-500 transition-all duration-300 cursor-pointer'
                     onClick={() => setEditable(!editable)}
                 />
                 <BiTrash
-                    className='text-3xl fill-red-500 cursor-pointer'
+                    className='text-3xl fill-red-400 hover:fill-red-500 transition-all duration-300 cursor-pointer'
                     onClick={confirmation}
                 />
             </div>
